Rename Runtime Auth to API Runtime Auth in director integration tests (#329)
