+++ chart/compass/values.yaml
@@ -10,7 +10,7 @@ global:
       version: "PR-247"
     director:
       dir: pr/
-      version: "PR-328"
+      version: "PR-329"
     gateway:
       dir: pr/
       version: "PR-251"
@@ -31,7 +31,7 @@ global:
     tests:
       e2e:
         dir: pr/
-        version: "PR-341"
+        version: "PR-329"
       connector:
         dir: pr/
         version: "PR-335"
