+++ examples/README.md
@@ -19,12 +19,12 @@ Below you can find a list of mock GraphQL queries and mutations:
 - [delete document](./delete-document.graphql)
 - [delete label definition](./delete-label-definition.graphql)
 - [delete runtime](./delete-runtime.graphql)
+- [query api runtime auths](./query-api-runtime-auths.graphql)
 - [query application](./query-application.graphql)
 - [query applications for runtime](./query-applications-for-runtime.graphql)
 - [query applications with label filter](./query-applications-with-label-filter.graphql)
 - [query applications](./query-applications.graphql)
 - [query label definition](./query-label-definition.graphql)
-- [query runtime auths](./query-runtime-auths.graphql)
 - [query runtime](./query-runtime.graphql)
 - [query runtimes with label filter](./query-runtimes-with-label-filter.graphql)
 - [query runtimes with pagination](./query-runtimes-with-pagination.graphql)
